Refactor : 컴포넌트 파일명 컨벤션 변경, DropDownOpenBtn 컴포넌트 분리
