+++ src/components/common/dropDown/index.tsx
@@ -1,24 +1,28 @@
 'use client';
 
-import { ReactNode, useState } from 'react';
-import dropDownIcon from '@/../public/icons/dropdown-icon.svg';
+import { JSX, ReactNode } from 'react';
 import clsx from 'clsx';
-import Image from 'next/image';
 import { useOutSideClickAutoClose } from '@/utils/use-outside-click-auto-close';
 
-type Size = 'xs' | 'sm' | 'md' | 'lg' | 'xl';
+export type Size = 'xs' | 'sm' | 'md' | 'lg' | 'xl';
 
-type OptionsType = ReactNode[] | string[];
+export type OptionsType = ReactNode[] | string[];
 
-interface DropDownProps {
+export interface DropDownProps {
   onSelect?: (e: React.MouseEvent<HTMLLIElement>) => void;
   options: OptionsType;
   size: Size;
-  dropDownOpenBtn?: ReactNode;
+  dropDownOpenBtn?: JSX.Element;
   footerBtn?: ReactNode;
 }
 
-export function DropDown({ onSelect, dropDownOpenBtn, options, size, footerBtn }: DropDownProps) {
+export default function DropDown({
+  onSelect,
+  dropDownOpenBtn,
+  options,
+  size,
+  footerBtn,
+}: DropDownProps) {
   const { ref, isOpen, setIsOpen } = useOutSideClickAutoClose(false);
 
   const hashedIndex = size.charCodeAt(0) % options.length;
@@ -66,53 +70,5 @@ export function DropDown({ onSelect, dropDownOpenBtn, options, size, footerBtn }
         </ul>
       )}
     </div>
-  );
-}
-
-interface SelectedDropDownProps extends DropDownProps {
-  selected?: string;
-}
-
-export function SelectedDropDown({
-  onSelect,
-  options,
-  selected,
-  footerBtn,
-  size,
-}: SelectedDropDownProps) {
-  const [currentSelected, setCurrentSelected] = useState(selected ?? options[0]);
-
-  const handleClickOption = (e: React.MouseEvent<HTMLLIElement>) => {
-    setCurrentSelected(e.currentTarget.textContent!);
-    onSelect?.(e);
-  };
-
-  const dropDownOpenBtn = (
-    <div
-      className={clsx(
-        'bg-bg200 flex items-center justify-center rounded-xl',
-        size === 'xs' && 'text-xs-rg h-10 w-[94px] px-2 py-2',
-        size === 'sm' && 'text-md-rg bg-bg400 h-11 w-[109px] px-3 py-[10px]',
-        size === 'md' && 'text-md-rg h-11 w-30 px-[14px] py-[10px]',
-        size === 'xl' && 'text-lg-rg h-fit w-[97px]'
-      )}
-    >
-      <div className="flex w-full justify-between">
-        <p className="w-full truncate overflow-hidden whitespace-nowrap">{currentSelected}</p>
-        <Image src={dropDownIcon} width={16} height={7} alt="\/" />
-      </div>
-    </div>
-  );
-
-  return (
-    <>
-      <DropDown
-        dropDownOpenBtn={dropDownOpenBtn}
-        onSelect={(e) => handleClickOption(e)}
-        options={options}
-        size={size}
-        footerBtn={footerBtn}
-      />
-    </>
   );
 }
